+++ admin/src/app/app-routing.module.ts
@@ -3,12 +3,15 @@ import { RouterModule, Routes } from '@angular/router';
 import { TripListComponent } from './trips/trip-list.component';
 import { TripAddComponent } from './trips/trip-add.component';
 import { TripEditComponent } from './trips/trip-edit.component';
+import { LoginComponent } from './auth/login.component';
+import { AuthGuard } from './auth/auth.guard';
 
 const routes: Routes = [
   { path: '', redirectTo: '/trips', pathMatch: 'full' },
-  { path: 'trips', component: TripListComponent },
-  { path: 'trips/add', component: TripAddComponent },
-  { path: 'trips/:code/edit', component: TripEditComponent },
+  { path: 'login', component: LoginComponent },
+  { path: 'trips', component: TripListComponent, canActivate: [AuthGuard] },
+  { path: 'trips/add', component: TripAddComponent, canActivate: [AuthGuard] },
+  { path: 'trips/:code/edit', component: TripEditComponent, canActivate: [AuthGuard] },
   { path: '**', redirectTo: '/trips' }
 ];